+++ CHANGELOG.md
@@ -1,10 +1,28 @@
 ## lifebit-ai/cloudos-cli: changelog
 
-## v2.61.0 (2025-09-24)
+## v2.62.0 (2025-10-02)
 
 ### Feat
 
 - Adds job cost details in stdout, csv and json formats
+
+## v2.61.2 (2025-09-25)
+
+### Patch
+
+- Adds accelerate saving results option for bash job and array-job
+
+## v2.61.1 (2025-09-25)
+
+### Patch
+
+- Fixes logging
+
+## v2.61.0 (2025-09-24)
+
+### Feat
+
+- Adds accelerate saving results option
 
 ## v2.60.5 (2025-09-22)